Fixed bot utils circular imports
