+++ packages/server/src/agent/websockets.test.ts
@@ -5,9 +5,10 @@ import express from 'express';
 import { Server } from 'http';
 import request from 'superwstest';
 import { initApp, shutdownApp } from '../app';
+import * as executeBotModule from '../bots/execute';
+import { BotExecutionResult } from '../bots/types';
 import { loadTestConfig } from '../config/loader';
 import { MedplumServerConfig } from '../config/types';
-import * as executeBotModule from '../fhir/operations/execute';
 import { getRedis } from '../redis';
 import { initTestAuth } from '../test.setup';
 import { AgentConnectionState, AgentInfo } from './utils';
@@ -653,7 +654,7 @@ describe('Agent WebSockets', () => {
                 '    at async exports.handler (/var/task/index.js:24:18)',
               ],
             },
-          }) satisfies executeBotModule.BotExecutionResult
+          }) satisfies BotExecutionResult
       );
 
       await request(server)
