+++ packages/server/src/agent/websockets.ts
@@ -13,8 +13,8 @@ import { Redis } from 'ioredis';
 import { AsyncLocalStorage } from 'node:async_hooks';
 import { IncomingMessage } from 'node:http';
 import ws from 'ws';
+import { executeBot } from '../bots/execute';
 import { getRepoForLogin } from '../fhir/accesspolicy';
-import { executeBot } from '../fhir/operations/execute';
 import { heartbeat } from '../heartbeat';
 import { globalLogger } from '../logger';
 import { getLoginForAccessToken } from '../oauth/utils';
